move not mojo specific code to common-module
