+++ maven/src/main/java/pl/project13/maven/git/GitCommitIdMojo.java
@@ -47,6 +47,7 @@ import org.eclipse.jgit.storage.file.FileRepositoryBuilder;
 
 import pl.project13.jgit.DescribeCommand;
 import pl.project13.jgit.DescribeResult;
+import pl.project13.maven.git.GitDataLoader.GitDataLoaderBuilder;
 import pl.project13.maven.git.log.LoggerBridge;
 import pl.project13.maven.git.log.MavenLoggerBridge;
 
@@ -72,20 +73,20 @@ import com.google.common.io.Files;
 @SuppressWarnings({"JavaDoc"})
 public class GitCommitIdMojo extends AbstractMojo {
 
-  // these properties will be exposed to maven
-  public static final String BRANCH = "branch";
-  public static final String COMMIT_ID = "commit.id";
-  public static final String COMMIT_ID_ABBREV = "commit.id.abbrev";
-  public static final String COMMIT_DESCRIBE = "commit.id.describe";
-  public static final String BUILD_AUTHOR_NAME = "build.user.name";
-  public static final String BUILD_AUTHOR_EMAIL = "build.user.email";
+//  // these properties will be exposed to maven
+//  public static final String BRANCH = "branch";
+//  public static final String COMMIT_ID = "commit.id";
+//  public static final String COMMIT_ID_ABBREV = "commit.id.abbrev";
+//  public static final String COMMIT_DESCRIBE = "commit.id.describe";
+//  public static final String BUILD_AUTHOR_NAME = "build.user.name";
+//  public static final String BUILD_AUTHOR_EMAIL = "build.user.email";
   public static final String BUILD_TIME = "build.time";
-  public static final String COMMIT_AUTHOR_NAME = "commit.user.name";
-  public static final String COMMIT_AUTHOR_EMAIL = "commit.user.email";
-  public static final String COMMIT_MESSAGE_FULL = "commit.message.full";
-  public static final String COMMIT_MESSAGE_SHORT = "commit.message.short";
-  public static final String COMMIT_TIME = "commit.time";
-  public static final String REMOTE_ORIGIN_URL = "remote.origin.url";
+//  public static final String COMMIT_AUTHOR_NAME = "commit.user.name";
+//  public static final String COMMIT_AUTHOR_EMAIL = "commit.user.email";
+//  public static final String COMMIT_MESSAGE_FULL = "commit.message.full";
+//  public static final String COMMIT_MESSAGE_SHORT = "commit.message.short";
+//  public static final String COMMIT_TIME = "commit.time";
+//  public static final String REMOTE_ORIGIN_URL = "remote.origin.url";
 
   /**
    * The maven project.
@@ -315,9 +316,29 @@ public class GitCommitIdMojo extends AbstractMojo {
       properties = initProperties();
       prefixDot = prefix + ".";
 
-      loadGitData(properties);
-      filterNot(properties, excludeProperties);
-      loadBuildTimeData(properties);
+//      loadGitData(properties);
+
+	  GitDataLoader gitDataLoader = new GitDataLoaderBuilder()
+	  								.withWorkingTreeDirectory(dotGitDirectory.getParentFile())
+	  								.withAbbrevationLength(abbrevLength)
+	  								.withDateFormat(dateFormat)
+	  								.withGitDescribe(gitDescribe)
+	  								.withPrefixDot(prefixDot)
+	  								.withLoggerBridge(loggerBridge)
+	  								.build();
+      
+      try {
+		gitDataLoader.loadGitData(properties);
+	} catch (GitDataLoaderException e) {
+		throw new MojoExecutionException(e.getMessage(), e);
+	}
+      
+//      filterNot(properties, excludeProperties);
+      PropertiesExcludeFilter.filterNot(properties, excludeProperties);
+      
+      // now in GitDataLoader
+//      loadBuildTimeData(properties);
+      
       logProperties(properties);
 
       if (generateGitPropertiesFile) {
@@ -333,29 +354,29 @@ public class GitCommitIdMojo extends AbstractMojo {
 
   }
 
-  private void filterNot(Properties properties, @Nullable List<String> exclusions) {
-    if (exclusions == null)
-      return;
-
-    List<Predicate<CharSequence>> excludePredicates = Lists.transform(exclusions, new Function<String, Predicate<CharSequence>>() {
-      @Override
-      public Predicate<CharSequence> apply(String exclude) {
-        return Predicates.containsPattern(exclude);
-      }
-    });
-
-    Predicate<CharSequence> shouldExclude = Predicates.alwaysFalse();
-    for (Predicate<CharSequence> predicate : excludePredicates) {
-      shouldExclude = Predicates.or(shouldExclude, predicate);
-    }
-
-    for (String key : properties.stringPropertyNames()) {
-      if (shouldExclude.apply(key)) {
-        System.out.println("shouldExclude.apply(" + key +") = " + shouldExclude.apply(key));
-        properties.remove(key);
-      }
-    }
-  }
+//  private void filterNot(Properties properties, @Nullable List<String> exclusions) {
+//    if (exclusions == null)
+//      return;
+//
+//    List<Predicate<CharSequence>> excludePredicates = Lists.transform(exclusions, new Function<String, Predicate<CharSequence>>() {
+//      @Override
+//      public Predicate<CharSequence> apply(String exclude) {
+//        return Predicates.containsPattern(exclude);
+//      }
+//    });
+//
+//    Predicate<CharSequence> shouldExclude = Predicates.alwaysFalse();
+//    for (Predicate<CharSequence> predicate : excludePredicates) {
+//      shouldExclude = Predicates.or(shouldExclude, predicate);
+//    }
+//
+//    for (String key : properties.stringPropertyNames()) {
+//      if (shouldExclude.apply(key)) {
+//        System.out.println("shouldExclude.apply(" + key +") = " + shouldExclude.apply(key));
+//        properties.remove(key);
+//      }
+//    }
+//  }
 
   /**
    * Reacts to an exception based on the {@code failOnUnableToExtractRepoInfo} setting.
@@ -423,112 +444,112 @@ public class GitCommitIdMojo extends AbstractMojo {
     return keyString.startsWith(prefixDot);
   }
 
-  void loadBuildTimeData(@Nonnull Properties properties) {
-    Date commitDate = new Date();
-    SimpleDateFormat smf = new SimpleDateFormat(dateFormat);
-    put(properties, BUILD_TIME, smf.format(commitDate));
-  }
+//  void loadBuildTimeData(@Nonnull Properties properties) {
+//    Date commitDate = new Date();
+//    SimpleDateFormat smf = new SimpleDateFormat(dateFormat);
+//    put(properties, BUILD_TIME, smf.format(commitDate));
+//  }
 
-  void loadGitData(@Nonnull Properties properties) throws IOException, MojoExecutionException {
-    Repository git = getGitRepository();
-    ObjectReader objectReader = git.newObjectReader();
+//  void loadGitData(@Nonnull Properties properties) throws IOException, MojoExecutionException {
+//    Repository git = getGitRepository();
+//    ObjectReader objectReader = git.newObjectReader();
+//
+//    // git.user.name
+//    String userName = git.getConfig().getString("user", null, "name");
+//    put(properties, BUILD_AUTHOR_NAME, userName);
+//
+//    // git.user.email
+//    String userEmail = git.getConfig().getString("user", null, "email");
+//    put(properties, BUILD_AUTHOR_EMAIL, userEmail);
+//
+//    // more details parsed out bellow
+//    Ref HEAD = git.getRef(Constants.HEAD);
+//    if (HEAD == null) {
+//      throw new MojoExecutionException("Could not get HEAD Ref, are you sure you've set the dotGitDirectory property of this plugin to a valid path?");
+//    }
+//    RevWalk revWalk = new RevWalk(git);
+//    RevCommit headCommit = revWalk.parseCommit(HEAD.getObjectId());
+//    revWalk.markStart(headCommit);
+//
+//    try {
+//      // git.branch
+//      String branch = determineBranchName(git, System.getenv());
+//      put(properties, BRANCH, branch);
+//
+//      // git.commit.id.describe
+//      maybePutGitDescribe(properties, git);
+//
+//      // git.commit.id
+//      put(properties, COMMIT_ID, headCommit.getName());
+//
+//      // git.commit.id.abbrev
+//      putAbbrevCommitId(objectReader, properties, headCommit, abbrevLength);
+//
+//      // git.commit.author.name
+//      String commitAuthor = headCommit.getAuthorIdent().getName();
+//      put(properties, COMMIT_AUTHOR_NAME, commitAuthor);
+//
+//      // git.commit.author.email
+//      String commitEmail = headCommit.getAuthorIdent().getEmailAddress();
+//      put(properties, COMMIT_AUTHOR_EMAIL, commitEmail);
+//
+//      // git commit.message.full
+//      String fullMessage = headCommit.getFullMessage();
+//      put(properties, COMMIT_MESSAGE_FULL, fullMessage);
+//
+//      // git commit.message.short
+//      String shortMessage = headCommit.getShortMessage();
+//      put(properties, COMMIT_MESSAGE_SHORT, shortMessage);
+//
+//      long timeSinceEpoch = headCommit.getCommitTime();
+//      Date commitDate = new Date(timeSinceEpoch * 1000); // git is "by sec" and java is "by ms"
+//      SimpleDateFormat smf = new SimpleDateFormat(dateFormat);
+//      put(properties, COMMIT_TIME, smf.format(commitDate));
+//
+//      // git remote.origin.url
+//      String remoteOriginUrl = git.getConfig().getString("remote", "origin", "url");
+//      put(properties, REMOTE_ORIGIN_URL, remoteOriginUrl);
+//    } finally {
+//      revWalk.dispose();
+//    }
+//  }
 
-    // git.user.name
-    String userName = git.getConfig().getString("user", null, "name");
-    put(properties, BUILD_AUTHOR_NAME, userName);
+//  private void putAbbrevCommitId(ObjectReader objectReader, Properties properties, RevCommit headCommit, int abbrevLength) throws MojoExecutionException {
+//    if (abbrevLength < 2 || abbrevLength > 40) {
+//      throw new MojoExecutionException("Abbreviated commit id lenght must be between 2 and 40, inclusive! Was [%s]. ".codePointBefore(abbrevLength) +
+//                                           "Please fix your configuration (the <abbrevLength/> element).");
+//    }
+//
+//    try {
+//      AbbreviatedObjectId abbreviatedObjectId = objectReader.abbreviate(headCommit, abbrevLength);
+//      put(properties, COMMIT_ID_ABBREV, abbreviatedObjectId.name());
+//    } catch (IOException e) {
+//      throw new MojoExecutionException("Unable to abbreviate commit id! " +
+//                                           "You may want to investigate the <abbrevLength/> element in your configuration.", e);
+//    }
+//  }
 
-    // git.user.email
-    String userEmail = git.getConfig().getString("user", null, "email");
-    put(properties, BUILD_AUTHOR_EMAIL, userEmail);
-
-    // more details parsed out bellow
-    Ref HEAD = git.getRef(Constants.HEAD);
-    if (HEAD == null) {
-      throw new MojoExecutionException("Could not get HEAD Ref, are you sure you've set the dotGitDirectory property of this plugin to a valid path?");
-    }
-    RevWalk revWalk = new RevWalk(git);
-    RevCommit headCommit = revWalk.parseCommit(HEAD.getObjectId());
-    revWalk.markStart(headCommit);
-
-    try {
-      // git.branch
-      String branch = determineBranchName(git, System.getenv());
-      put(properties, BRANCH, branch);
-
-      // git.commit.id.describe
-      maybePutGitDescribe(properties, git);
-
-      // git.commit.id
-      put(properties, COMMIT_ID, headCommit.getName());
-
-      // git.commit.id.abbrev
-      putAbbrevCommitId(objectReader, properties, headCommit, abbrevLength);
-
-      // git.commit.author.name
-      String commitAuthor = headCommit.getAuthorIdent().getName();
-      put(properties, COMMIT_AUTHOR_NAME, commitAuthor);
-
-      // git.commit.author.email
-      String commitEmail = headCommit.getAuthorIdent().getEmailAddress();
-      put(properties, COMMIT_AUTHOR_EMAIL, commitEmail);
-
-      // git commit.message.full
-      String fullMessage = headCommit.getFullMessage();
-      put(properties, COMMIT_MESSAGE_FULL, fullMessage);
-
-      // git commit.message.short
-      String shortMessage = headCommit.getShortMessage();
-      put(properties, COMMIT_MESSAGE_SHORT, shortMessage);
-
-      long timeSinceEpoch = headCommit.getCommitTime();
-      Date commitDate = new Date(timeSinceEpoch * 1000); // git is "by sec" and java is "by ms"
-      SimpleDateFormat smf = new SimpleDateFormat(dateFormat);
-      put(properties, COMMIT_TIME, smf.format(commitDate));
-
-      // git remote.origin.url
-      String remoteOriginUrl = git.getConfig().getString("remote", "origin", "url");
-      put(properties, REMOTE_ORIGIN_URL, remoteOriginUrl);
-    } finally {
-      revWalk.dispose();
-    }
-  }
-
-  private void putAbbrevCommitId(ObjectReader objectReader, Properties properties, RevCommit headCommit, int abbrevLength) throws MojoExecutionException {
-    if (abbrevLength < 2 || abbrevLength > 40) {
-      throw new MojoExecutionException("Abbreviated commit id lenght must be between 2 and 40, inclusive! Was [%s]. ".codePointBefore(abbrevLength) +
-                                           "Please fix your configuration (the <abbrevLength/> element).");
-    }
-
-    try {
-      AbbreviatedObjectId abbreviatedObjectId = objectReader.abbreviate(headCommit, abbrevLength);
-      put(properties, COMMIT_ID_ABBREV, abbreviatedObjectId.name());
-    } catch (IOException e) {
-      throw new MojoExecutionException("Unable to abbreviate commit id! " +
-                                           "You may want to investigate the <abbrevLength/> element in your configuration.", e);
-    }
-  }
-
-  void maybePutGitDescribe(@Nonnull Properties properties, @Nonnull Repository repository) throws MojoExecutionException {
-    if (gitDescribe == null || !gitDescribe.isSkip()) {
-      putGitDescribe(properties, repository);
-    }
-  }
-
-  @VisibleForTesting
-  void putGitDescribe(@Nonnull Properties properties, @Nonnull Repository repository) throws MojoExecutionException {
-    try {
-      DescribeResult describeResult = DescribeCommand
-          .on(repository)
-          .withLoggerBridge(loggerBridge)
-          .setVerbose(verbose)
-          .apply(gitDescribe)
-          .call();
-
-      put(properties, COMMIT_DESCRIBE, describeResult.toString());
-    } catch (GitAPIException ex) {
-      throw new MojoExecutionException("Unable to obtain git.commit.id.describe information", ex);
-    }
-  }
+//  void maybePutGitDescribe(@Nonnull Properties properties, @Nonnull Repository repository) throws MojoExecutionException {
+//    if (gitDescribe == null || !gitDescribe.isSkip()) {
+//      putGitDescribe(properties, repository);
+//    }
+//  }
+//
+//  @VisibleForTesting
+//  void putGitDescribe(@Nonnull Properties properties, @Nonnull Repository repository) throws MojoExecutionException {
+//    try {
+//      DescribeResult describeResult = DescribeCommand
+//          .on(repository)
+//          .withLoggerBridge(loggerBridge)
+//          .setVerbose(verbose)
+//          .apply(gitDescribe)
+//          .call();
+//
+//      put(properties, COMMIT_DESCRIBE, describeResult.toString());
+//    } catch (GitAPIException ex) {
+//      throw new MojoExecutionException("Unable to obtain git.commit.id.describe information", ex);
+//    }
+//  }
 
   static int counter;
 
@@ -559,44 +580,44 @@ public class GitCommitIdMojo extends AbstractMojo {
     return project.getPackaging().equalsIgnoreCase("pom");
   }
 
-  @Nonnull
-  private Repository getGitRepository() throws MojoExecutionException {
-    Repository repository;
+//  @Nonnull
+//  private Repository getGitRepository() throws MojoExecutionException {
+//    Repository repository;
+//
+//    FileRepositoryBuilder repositoryBuilder = new FileRepositoryBuilder();
+//    try {
+//      repository = repositoryBuilder
+//          .setGitDir(dotGitDirectory)
+//          .readEnvironment() // scan environment GIT_* variables
+//          .findGitDir() // scan up the file system tree
+//          .build();
+//    } catch (IOException e) {
+//      throw new MojoExecutionException("Could not initialize repository...", e);
+//    }
+//
+//    if (repository == null) {
+//      throw new MojoExecutionException("Could not create git repository. Are you sure '" + dotGitDirectory + "' is the valid Git root for your project?");
+//    }
+//
+//    return repository;
+//  }
+//
+//  private void put(@Nonnull Properties properties, String key, String value) {
+//    putWithoutPrefix(properties, prefixDot + key, value);
+//  }
 
-    FileRepositoryBuilder repositoryBuilder = new FileRepositoryBuilder();
-    try {
-      repository = repositoryBuilder
-          .setGitDir(dotGitDirectory)
-          .readEnvironment() // scan environment GIT_* variables
-          .findGitDir() // scan up the file system tree
-          .build();
-    } catch (IOException e) {
-      throw new MojoExecutionException("Could not initialize repository...", e);
-    }
+//  private void putWithoutPrefix(@Nonnull Properties properties, String key, String value) {
+//    if (!isNotEmpty(value)) {
+//      value = "Unknown";
+//    }
+//
+//    log(key, value);
+//    properties.put(key, value);
+//  }
 
-    if (repository == null) {
-      throw new MojoExecutionException("Could not create git repository. Are you sure '" + dotGitDirectory + "' is the valid Git root for your project?");
-    }
-
-    return repository;
-  }
-
-  private void put(@Nonnull Properties properties, String key, String value) {
-    putWithoutPrefix(properties, prefixDot + key, value);
-  }
-
-  private void putWithoutPrefix(@Nonnull Properties properties, String key, String value) {
-    if (!isNotEmpty(value)) {
-      value = "Unknown";
-    }
-
-    log(key, value);
-    properties.put(key, value);
-  }
-
-  private boolean isNotEmpty(@Nullable String value) {
-    return null != value && !" ".equals(value.trim().replaceAll(" ", ""));
-  }
+//  private boolean isNotEmpty(@Nullable String value) {
+//    return null != value && !" ".equals(value.trim().replaceAll(" ", ""));
+//  }
 
   void log(String... parts) {
     loggerBridge.log((Object[]) parts);
@@ -610,49 +631,49 @@ public class GitCommitIdMojo extends AbstractMojo {
     return !directoryExists(fileLocation);
   }
 
-  /**
-   * If running within Jenkins/Hudosn, honor the branch name passed via GIT_BRANCH env var.  This
-   * is necessary because Jenkins/Hudson alwways invoke build in a detached head state.
-   *
-   * @param git
-   * @param env
-   * @return results of git.getBranch() or, if in Jenkins/Hudson, value of GIT_BRANCH
-   */
-  protected String determineBranchName(Repository git, Map<String, String> env) throws IOException {
-    if (runningOnBuildServer(env)) {
-      return determineBranchNameOnBuildServer(git, env);
-    } else {
-      return git.getBranch();
-    }
-  }
+//  /**
+//   * If running within Jenkins/Hudosn, honor the branch name passed via GIT_BRANCH env var.  This
+//   * is necessary because Jenkins/Hudson alwways invoke build in a detached head state.
+//   *
+//   * @param git
+//   * @param env
+//   * @return results of git.getBranch() or, if in Jenkins/Hudson, value of GIT_BRANCH
+//   */
+//  protected String determineBranchName(Repository git, Map<String, String> env) throws IOException {
+//    if (runningOnBuildServer(env)) {
+//      return determineBranchNameOnBuildServer(git, env);
+//    } else {
+//      return git.getBranch();
+//    }
+//  }
 
-  /**
-   * Is "Jenkins aware", and prefers {@code GIT_BRANCH} to getting the branch via git if that enviroment variable is set.
-   * The {@GIT_BRANCH} variable is set by Jenkins/Hudson when put in detached HEAD state, but it still knows which branch was cloned.
-   */
-  protected String determineBranchNameOnBuildServer(Repository git, Map<String, String> env) throws IOException {
-    String enviromentBasedBranch = env.get("GIT_BRANCH");
-    if(isNullOrEmpty(enviromentBasedBranch)) {
-      log("Detected that running on CI enviroment, but using repository branch, no GIT_BRANCH detected.");
-      return git.getBranch();
-    }else {
-      log("Using environment variable based branch name.", "GIT_BRANCH =", enviromentBasedBranch);
-      return enviromentBasedBranch;
-    }
-  }
-
-  /**
-   * Detects if we're running on Jenkins or Hudson, based on expected env variables.
-   * <p/>
-   * TODO: How can we detect Bamboo, TeamCity etc? Pull requests welcome.
-   *
-   * @return true if running
-   * @see <a href="https://wiki.jenkins-ci.org/display/JENKINS/Building+a+software+project#Buildingasoftwareproject-JenkinsSetEnvironmentVariables">JenkinsSetEnvironmentVariables</a>
-   * @param env
-   */
-  private boolean runningOnBuildServer(Map<String, String> env) {
-    return env.containsKey("HUDSON_URL") || env.containsKey("JENKINS_URL");
-  }
+//  /**
+//   * Is "Jenkins aware", and prefers {@code GIT_BRANCH} to getting the branch via git if that enviroment variable is set.
+//   * The {@GIT_BRANCH} variable is set by Jenkins/Hudson when put in detached HEAD state, but it still knows which branch was cloned.
+//   */
+//  protected String determineBranchNameOnBuildServer(Repository git, Map<String, String> env) throws IOException {
+//    String enviromentBasedBranch = env.get("GIT_BRANCH");
+//    if(isNullOrEmpty(enviromentBasedBranch)) {
+//      log("Detected that running on CI enviroment, but using repository branch, no GIT_BRANCH detected.");
+//      return git.getBranch();
+//    }else {
+//      log("Using environment variable based branch name.", "GIT_BRANCH =", enviromentBasedBranch);
+//      return enviromentBasedBranch;
+//    }
+//  }
+//
+//  /**
+//   * Detects if we're running on Jenkins or Hudson, based on expected env variables.
+//   * <p/>
+//   * TODO: How can we detect Bamboo, TeamCity etc? Pull requests welcome.
+//   *
+//   * @return true if running
+//   * @see <a href="https://wiki.jenkins-ci.org/display/JENKINS/Building+a+software+project#Buildingasoftwareproject-JenkinsSetEnvironmentVariables">JenkinsSetEnvironmentVariables</a>
+//   * @param env
+//   */
+//  private boolean runningOnBuildServer(Map<String, String> env) {
+//    return env.containsKey("HUDSON_URL") || env.containsKey("JENKINS_URL");
+//  }
 
   // SETTERS FOR TESTS ----------------------------------------------------
